requirements.txt
Django==1.6.5
Djangorestframework==2.4.3
Markdown==2.5.1